+++ resources/views/layouts/admin.blade.php
@@ -100,77 +100,8 @@
 <!-- jvectormap  -->
 <script src="{{ asset("js/jquery-jvectormap-1.2.2.min.js")}}"></script>
 <script src="{{ asset("js/jquery-jvectormap-world-mill-en.js")}}"></script>
+
+
+@stack('scripts')
 </body>
 </html>
-<script>  
-    $(document).ready(function() {
-       //Initialize Select2 Elements
-       $('.select2').select2();       
-
-       $(".option").click(function(){   
-
-          $(".incidenciasform").toggle("slide");         
-       });
-   
-       //Datemask dd/mm/yyyy
-       $('#datemask').inputmask('dd/mm/yyyy', { 'placeholder': 'dd/mm/yyyy' })
-       //Datemask2 mm/dd/yyyy
-       $('#datemask2').inputmask('mm/dd/yyyy', { 'placeholder': 'mm/dd/yyyy' })
-       //Money Euro
-       $('[data-mask]').inputmask()
-   
-       //Date range picker
-       $('#reservation').daterangepicker()
-       //Date range picker with time picker
-       $('#reservationtime').daterangepicker({ timePicker: true, timePickerIncrement: 30, format: 'MM/DD/YYYY h:mm A' })
-       //Date range as a button
-       $('#daterange-btn').daterangepicker(
-         {
-           ranges   : {
-             'Today'       : [moment(), moment()],
-             'Yesterday'   : [moment().subtract(1, 'days'), moment().subtract(1, 'days')],
-             'Last 7 Days' : [moment().subtract(6, 'days'), moment()],
-             'Last 30 Days': [moment().subtract(29, 'days'), moment()],
-             'This Month'  : [moment().startOf('month'), moment().endOf('month')],
-             'Last Month'  : [moment().subtract(1, 'month').startOf('month'), moment().subtract(1, 'month').endOf('month')]
-           },
-           startDate: moment().subtract(29, 'days'),
-           endDate  : moment()
-         },
-         function (start, end) {
-           $('#daterange-btn span').html(start.format('MMMM D, YYYY') + ' - ' + end.format('MMMM D, YYYY'))
-         }
-       )
-   
-       //Date picker
-       $('#datepicker').datepicker({
-         autoclose: true
-       })
-   
-       //iCheck for checkbox and radio inputs
-       $('input[type="checkbox"].minimal, input[type="radio"].minimal').iCheck({
-         checkboxClass: 'icheckbox_minimal-blue',
-         radioClass   : 'iradio_minimal-blue'
-       })
-       //Red color scheme for iCheck
-       $('input[type="checkbox"].minimal-red, input[type="radio"].minimal-red').iCheck({
-         checkboxClass: 'icheckbox_minimal-red',
-         radioClass   : 'iradio_minimal-red'
-       })
-       //Flat red color scheme for iCheck
-       $('input[type="checkbox"].flat-red, input[type="radio"].flat-red').iCheck({
-         checkboxClass: 'icheckbox_flat-green',
-         radioClass   : 'iradio_flat-green'
-       })
-   
-       //Colorpicker
-       $('.my-colorpicker1').colorpicker()
-       //color picker with addon
-       $('.my-colorpicker2').colorpicker()
-   
-       //Timepicker
-       $('.timepicker').timepicker({
-         showInputs: false
-       })
-     })
-   </script>
