+++ resources/views/layouts/footer.blade.php
@@ -1,4 +1,6 @@
 
   <footer class="main-footer">       
-        <strong>Copyright &copy; 2019 <a href="https://apimaker.com.ve">Apimaker Group</a>.</strong> Todos los Derechos Reservados.
+        <strong>Copyright &copy; <script>
+            document.write(new Date().getFullYear())
+          </script> <a href="https://apimaker.com.ve">Apimaker Group</a>.</strong> Todos los Derechos Reservados.
       </footer>
