+++ resources/views/layouts/sidebar.blade.php
@@ -8,7 +8,7 @@
               <img src="{{ asset("dist/img/user2-160x160.jpg")}}" class="img-circle" alt="User Image">
             </div>
             <div class="pull-left info">
-              <p>Alexander Pierce</p>
+            <p>{{auth()->user()->name}}</p>
               <a href="#"><i class="fa fa-circle text-success"></i> Online</a>
             </div>
           </div>
